+++ CrocoDash/data_access/datasets/glorys.py
@@ -10,7 +10,7 @@ from CrocoDash.rm6 import regional_mom6 as rm6
 from pathlib import Path
 from CrocoDash.data_access.utils import fill_template
 import pandas as pd
-from .utils import setup_logger
+from ..utils import setup_logger
 
 logger = setup_logger(__name__)
