+++ CrocoDash/data_access/driver.py
@@ -47,7 +47,7 @@ class ProductFunctionRegistry:
             )
             try:
                 module = importlib.import_module(
-                    "." + submodule, package="CrocoDash.data_access"
+                    "." + submodule, package="CrocoDash.data_access.datasets"
                 )
                 func = getattr(module, func_name)
